+++ packages/components/elvia-components.config.js
@@ -156,7 +156,6 @@ module.exports = [
       { name: 'customError', type: 'string', propType: 'string | undefined' },
       { name: 'minDate', type: 'Date', propType: 'Date | undefined' },
       { name: 'maxDate', type: 'Date', propType: 'Date | undefined' },
-      { name: 'initialFocusedDate', type: 'Date', propType: 'Date | undefined' },
       { name: 'placeholder', type: 'string', propType: 'string | undefined' },
       { name: 'open', type: 'boolean', propType: 'boolean' },