+++ packages/components/components/elvis-datepicker/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 2.1.1 (25.01.22)
 
 - Remove fixed width
-- New props: initialFocusedDate, placeholder and open
+- New props: placeholder and open
 
 ## 2.1.0 (22.12.21)